contextforge-ai/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
pgvector>=0.3.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
torch>=2.3.0
transformers>=4.41.0
jinja2>=3.1.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
pyarrow>=16.0.0
python-multipart>=0.0.9
pyyaml>=6.0.0
streamlit>=1.35.0
pytest>=8.2.0
